\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap}
\usage{
  Bootstrap(x, Boots = 100, fn)
}
\arguments{
  \item{x}{a vector.}

  \item{Boots}{The number of bootstraps}

  \item{fn}{the function you want to bootstrap}
}
\description{
  Bootstrap
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

