\name{CompPack}
\alias{CompPack}
\title{A collection of Computational Statistic Functions.}
\arguments{
  \item{anneal}{Uphill search on function}

  \item{around}{Find the Values Around a Particular Value}

  \item{bootstrap}{Bootstrap a simple function}

  \item{cvscore}{Cross Validation for 1-5 order polynomials}

  \item{cvscore1}{Cross Validation for 1st order polynomials}

  \item{cvsscore2}{Cross Validation for 2nd order polynomials}
  
  \item{cvsscore3}{Cross Validation for 3rd order polynomials}
  
  \item{cvsscore4}{Cross Validation for 4th order polynomials}
  
  \item{cvsscore5}{Cross Validation for 5th order polynomials}
  
  \item{d1f}{Take first derivative of a function}
  
  \item{df}{Take second derivative of a function}
  
  \item{em.mixnorm}{EM-Algorithm for Normal Distribution}
  
  \item{f.models}{Plot a linear, quadratic, cubic and 4th degree models for
  a 2D dataset.}
  
  \item{gold.sect}{Golden Section Search Optimization}
  
  \item{grid.sect}{Grid Section Search Optimization}
  
  \item{jackknife}{Resamples Data using the Jackknife Method}
  
  \item{log.lik}{Find the log likelihood of a normal distribution}
  
  \item{maxbound}{Maxbound Optimization}
  
  \item{mc}{Generate Markov Chains}
  
  \item{mixlike}{Compute the log likelihood function for the mixture of
  normal distributions}
  
  \item{mixnorm}{Create a Mixture of Normal Distributions}
  
  \item{nearest.loc}{Find Location of Nearest Value}
  
  \item{newton.raph}{Newton Raphson Alogorithm}
  
  \item{onedfunction}{Simple polynomial function}
  
  \item{plot.mc}{Plot Markov Chains}
  
  \item{pois.int}{Averages the length of an interval containing a
  specified time}
  
  \item{pois.proc}{Simulate a Poisson process of specified length and rate.}
  
  \item{pois.rate}{Average Time Between Events in a Poisson Process}
  
  \item{polar.rnorm}{Polar Random Normal}
  
  \item{rand.exp}{Generate Random Exponential Variables}
  
  \item{rand.mvrnorm}{Generate Multivariate Normal Distribution}
  
  \item{rand.norm}{Generate Random Normal Variables}
  
  \item{rand.unif}{Generate Random Uniform Variables}
  
  \item{rand.weib}{Generate Random Weibull Variables}
  
  \item{tpower.data}{Matrix of p-values from the tpower function}
  
  \item{twodfunction}{Simple 2D function for demonstrations}
  
  \item{update.ComPack}{Updates CompPack}
  
  \item{uphill.mixnorm}{uphill search on a mixture of normal distribution}
  
  \item{wich.hill}{Wichman Hill RNG}
}
\description{
  General anneal function, where we conduct a uphill search
  on the function to find maximum
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com} 
  
  Damon McCafferty \email{damon.mccafferty@economics.utah.edu}
  
  Katie Dodds \email{k.dodds@utah.edu}
  
  Ernest Chan \email{faiernest418@gmail.com}
}

