\name{f.models}
\alias{f.models}
\title{f models}
\usage{
  f.models(x, y)
}
\arguments{
  \item{x}{is the independent variable}

  \item{y}{is the dependent variable}
}
\description{
  Plots linear, quadratic, cubic and 4th degree models for
  a 2D dataset.
}
\examples{
attach(mtcars)
f.models(wt,mpg)
}
\author{
  Katie Dodds \email{k.dodds@utah.edu}
}

