\name{gold.sect}
\alias{gold.sect}
\title{Golden Section Search Optimization}
\usage{
  gold.sect(f, lo, hi, tol = 1e-07)
}
\arguments{
  \item{f}{The function to be optimized}

  \item{lo}{Lower estimate}

  \item{hi}{Upper estimate}

  \item{tol}{Level of tolerance desired for optimization}
}
\description{
  Golden Section Search Optimization
}

