\name{grid.sect}
\alias{grid.sect}
\title{Grid Section Search Optimization}
\usage{
  grid.sect(f, lo, hi)
}
\arguments{
  \item{f}{the function to be optimized.}

  \item{lo}{the lower bound.}

  \item{hi}{the upper bound.}
}
\description{
  A method used to locate the optimum value of a function.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

