\name{mvrand.norm}
\alias{mvrand.norm}
\title{Generate Multivariate Normal Distribution}
\usage{
  mvrand.norm(n, mu, sigma)
}
\arguments{
  \item{n}{the number of observations}

  \item{mu}{a vector of means}

  \item{sigma}{a variance-covariance matrix}
}
\description{
  A method for generating random multivariate normal data
  that is based on a cholesky decomposition.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

