\name{nearest.loc}
\alias{nearest.loc}
\title{Find Location of Nearest Value}
\usage{
  nearest.loc(x, value)
}
\arguments{
  \item{x}{a vector.}

  \item{value}{the value that you want to find.}
}
\description{
  Find the location of the nearest value to a number that
  you specify.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

