\name{nearest}
\alias{nearest}
\title{Find the Nearest Value}
\usage{
  nearest(x, value)
}
\arguments{
  \item{x}{a vector.}

  \item{value}{the value that you want to find.}
}
\description{
  Find the the nearest value to a number that you specify.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

