\name{polar.norm}
\alias{polar.norm}
\title{Generate A Random Normal Variable}
\usage{
  polar.norm(mu, sig)
}
\arguments{
  \item{mu}{mean of a normal distribution}

  \item{sig}{variance of a normal distribution}
}
\description{
  A method for generating a random normal variable based
  upon the polar method.
}
\author{
  Damon McCafferty
  \email{damon.mccafferty@economics.utah.edu}
}

