\name{polar.rnorm}
\alias{polar.rnorm}
\title{Polar Random Normal}
\usage{
  polar.rnorm(n = 100, mu = 0, sd = 1)
}
\arguments{
  \item{n}{is the sample size}

  \item{mu}{a vector of means}

  \item{sd}{a vector of standard deviations}
}
\description{
  Generates a sample of standard normals
}
\examples{
polar.rnorm(100)
}
\author{
  Katie Dodds \email{k.dodds@utah.edu} Damon McCafferty
  \email{damon.mccafferty@economics.utah.edu}
}

