\name{rand.exp}
\alias{rand.exp}
\title{Generate Random Exponential Variables}
\usage{
  rand.exp(n, rate = 1)
}
\arguments{
  \item{n}{Number of observations}

  \item{rate}{vector of rates}
}
\description{
  Generates a pseudo-random exponential variable using the
  Whichman-Hill pseudo-random number generator function
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

