\name{rand.norm}
\alias{rand.norm}
\title{Generate Random Normal Variables}
\usage{
  rand.norm(n, mean, sd)
}
\arguments{
  \item{n}{Number of observations.}

  \item{mean}{vector of means.}

  \item{sd}{vector of standard deviations.}
}
\description{
  Generates pseudo random variables from a normal
  distribution using the Wichmannhill.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

