\name{rand.unif}
\alias{rand.unif}
\title{Generate Random Uniform Variables}
\usage{
  rand.unif(n, min = 0, max = 1)
}
\arguments{
  \item{n}{Number of observations.}

  \item{min}{lower limit of the distribution.}

  \item{max}{upper limit of the distribution.}
}
\description{
  Generates pseudo-random variables from a Uniform
  distribution, using the Wichmannhill.
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

