\name{rand.weib}
\alias{rand.weib}
\title{Generate Random Weibull Variables}
\usage{
  rand.weib(n, beta, theta)
}
\arguments{
  \item{n}{the number of observations}

  \item{beta}{the specification of beta}

  \item{theta}{the specification of theta}
}
\description{
  Generate Random Weibull Variables
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

