\name{twodfunction}
\alias{twodfunction}
\title{twodfunction}
\usage{
  twodfunction(x, y)
}
\arguments{
  \item{x}{vector of value}

  \item{y}{vector of value}
}
\description{
  Generate cos(x) + cos(y) maxtrix
}
\examples{
twodfunction(1,1)
}
\author{
  Ernest Chan \email{faiernest@gmail.com}
}

