\name{around}
\alias{around}
\title{Find the Values Around a Particular Value}
\usage{
  around(x, value)
}
\arguments{
  \item{x}{a vector.}

  \item{value}{specified value}
}
\description{
  Find the location of values around a specified value
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

