\name{em.mixnorm}
\alias{em.mixnorm}
\title{EM-Algorithm for Normal Distribution}
\usage{
  em.mixnorm(x, k)
}
\arguments{
  \item{x}{the data.}

  \item{k}{estimate of the number of mixtures.}
}
\description{
  EM-Algorithm for Normal Distribution
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

