\name{mixnorm}
\alias{mixnorm}
\title{Create a Mixture of Normal Distributions}
\usage{
  mixnorm(n, p, mu, sd)
}
\arguments{
  \item{n}{number of observations. If \eqn{length(n) > 1},
  the length is taken to be the number required}

  \item{p}{vector of probabilities}

  \item{mu}{vector of means}

  \item{sd}{vector of standard deviations}
}
\description{
  Create a Mixture of Normal Distributions
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

