\name{pois.proc}
\alias{pois.proc}
\title{Generate a Poisson Process}
\usage{
  pois.proc(end = 200, rate = 1)
}
\arguments{
  \item{end}{the end time.}

  \item{rate}{the rate of occurrence.}
}
\description{
  Generate a Poisson Process
}
\examples{
pois.proc(200, 1)
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

