\name{rand.norm}
\alias{rand.norm}
\title{Generate Random Normal Variables}
\usage{
  rand.norm(n, mean, sd)
}
\arguments{
  \item{n}{Number of observations}

  \item{mean}{vector of means}

  \item{sd}{vector of standard deviations}
}
\description{
  Generate Random Normal Variables
}
\author{
  Tyler Hunt \email{tyler@psychoanalytix.com}
}

